+++ src/Geocoder/Provider/GoogleMaps.php
@@ -265,6 +265,7 @@ class GoogleMaps extends AbstractHttpProvider implements LocaleAwareProvider
                 break;
 
             case 'locality':
+            case 'postal_town':
                 $resultSet['locality'] = $values->long_name;
                 break;