+++ src/Geocoder/Provider/OpenCage.php
@@ -7,7 +7,6 @@
  *
  * @license    MIT License
  */
-
 namespace Geocoder\Provider;
 
 use Geocoder\Exception\InvalidCredentials;
@@ -53,7 +52,7 @@ class OpenCage extends AbstractHttpProvider implements LocaleAwareProvider
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function geocode($address)
     {
@@ -66,23 +65,23 @@ class OpenCage extends AbstractHttpProvider implements LocaleAwareProvider
             throw new UnsupportedOperation('The OpenCage provider does not support IP addresses, only street addresses.');
         }
 
-        $query = sprintf(self::GEOCODE_ENDPOINT_URL, $this->scheme, $this->apiKey, urlencode($address), $this->getLimit() );
+        $query = sprintf(self::GEOCODE_ENDPOINT_URL, $this->scheme, $this->apiKey, urlencode($address), $this->getLimit());
 
         return $this->executeQuery($query);
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function reverse($latitude, $longitude)
     {
-        $address = sprintf("%f, %f", $latitude, $longitude);
+        $address = sprintf('%f, %f', $latitude, $longitude);
 
         return $this->geocode($address);
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function getName()
     {
@@ -90,7 +89,8 @@ class OpenCage extends AbstractHttpProvider implements LocaleAwareProvider
     }
 
     /**
-     * @param string $query
+     * @param $query
+     * @return \Geocoder\Model\AddressCollection
      */
     private function executeQuery($query)
     {
@@ -106,7 +106,7 @@ class OpenCage extends AbstractHttpProvider implements LocaleAwareProvider
 
         $json = json_decode($content, true);
 
-        if (!isset($json['total_results']) || $json['total_results'] == 0 ) {
+        if (!isset($json['total_results']) || $json['total_results'] == 0) {
             throw new NoResult(sprintf('Could not find results for query "%s".', $query));
         }
 
@@ -122,9 +122,9 @@ class OpenCage extends AbstractHttpProvider implements LocaleAwareProvider
             if (isset($location['bounds'])) {
                 $bounds = [
                     'south' => $location['bounds']['southwest']['lat'],
-                    'west'  => $location['bounds']['southwest']['lng'],
+                    'west' => $location['bounds']['southwest']['lng'],
                     'north' => $location['bounds']['northeast']['lat'],
-                    'east'  => $location['bounds']['northeast']['lng'],
+                    'east' => $location['bounds']['northeast']['lng'],
                 ];
             }
 
@@ -138,21 +138,74 @@ class OpenCage extends AbstractHttpProvider implements LocaleAwareProvider
             }
 
             $results[] = array_merge($this->getDefaults(), array(
-                'latitude'     => $location['geometry']['lat'],
-                'longitude'    => $location['geometry']['lng'],
-                'bounds'       => $bounds ?: [],
+                'latitude' => $location['geometry']['lat'],
+                'longitude' => $location['geometry']['lng'],
+                'bounds' => $bounds ?: [],
                 'streetNumber' => isset($comp['house_number']) ? $comp['house_number'] : null,
-                'streetName'   => isset($comp['road']        ) ? $comp['road']         : null,
-                'subLocality'  => isset($comp['suburb']      ) ? $comp['suburb']       : null,
-                'locality'     => isset($comp['city']        ) ? $comp['city']         : null,
-                'postalCode'   => isset($comp['postcode']    ) ? $comp['postcode']     : null,
-                'adminLevels'  => $adminLevels,
-                'country'      => isset($comp['country']     ) ? $comp['country']      : null,
-                'countryCode'  => isset($comp['country_code']) ? strtoupper($comp['country_code']) : null,
-                'timezone'     => isset($location['annotations']['timezone']['name']) ? $location['annotations']['timezone']['name'] : null,
+                'streetName' => $this->guessStreetName($comp),
+                'subLocality' => $this->guessSubLocality($comp),
+                'locality' => $this->guessLocality($comp),
+                'postalCode' => isset($comp['postcode']) ? $comp['postcode']     : null,
+                'adminLevels' => $adminLevels,
+                'country' => isset($comp['country']) ? $comp['country']      : null,
+                'countryCode' => isset($comp['country_code']) ? strtoupper($comp['country_code']) : null,
+                'timezone' => isset($location['annotations']['timezone']['name']) ? $location['annotations']['timezone']['name'] : null,
             ));
         }
 
         return $this->returnResults($results);
+    }
+
+    /**
+     * @param array $components
+     *
+     * @return null|string
+     */
+    protected function guessLocality(array $components)
+    {
+        $localityKeys = array('city', 'town' , 'village', 'hamlet');
+
+        return $this->guessBestComponent($components, $localityKeys);
+    }
+
+    /**
+     * @param array $components
+     *
+     * @return null|string
+     */
+    protected function guessStreetName(array $components)
+    {
+        $streetNameKeys = array('road', 'street', 'street_name', 'residential');
+
+        return $this->guessBestComponent($components, $streetNameKeys);
+    }
+
+    /**
+     * @param array $components
+     *
+     * @return null|string
+     */
+    protected function guessSubLocality(array $components)
+    {
+        $subLocalityKeys = array('suburb', 'neighbourhood', 'city_district');
+
+        return $this->guessBestComponent($components, $subLocalityKeys);
+    }
+
+    /**
+     * @param array $components
+     * @param array $keys
+     *
+     * @return null|string
+     */
+    protected function guessBestComponent(array $components, array $keys)
+    {
+        foreach ($keys as $key) {
+            if (isset($components[$key]) && !empty($components[$key])) {
+                return $components[$key];
+            }
+        }
+
+        return null;
     }
 }